fix: fix bugs introduced in last commit
