+++ polus/_plugins/PolusComputeSchema.py
@@ -7,9 +7,10 @@ from __future__ import annotations
 from enum import Enum
 from typing import List, Optional, Union, Any
 
-from pydantic import BaseModel, Field, constr
+from pydantic import BaseModel, Field, constr, validator
 
 from ._io import IOBase, Version
+from ._utils import utils_cast_version
 
 
 class ConditionEntry(BaseModel):
@@ -141,6 +142,10 @@ class PluginSchema(BaseModel):
     ui: List[Union[PluginUIInput, PluginUIOutput]]
     version: Version
     website: Optional[str] = None
+    
+    @validator("version", pre=True)
+    def cast_version(cls, value):
+        return utils_cast_version(value)
 
 
 class CLTSchema(PluginSchema):
