+++ polus/_plugins/_utils.py
@@ -1,5 +1,11 @@
+from ._io import Version
 def name_cleaner(name: str) -> str:
     replace_chars = "()<>-_"
     for char in replace_chars:
         name = name.replace(char, " ")
     return name.title().replace(" ", "").replace("/", "_")
+
+def utils_cast_version(value):
+    if isinstance(value, dict):  # if init from a Version object
+        value = value["version"]
+    return Version(version=value)
